Backend/requirements.txt
pandas>=1.5
PyYAML>=6.0
geopandas>=0.13
fiona>=1.9
shapely>=2.0
rtree>=0.10
pyarrow>=9.0
scikit-learn>=1.1
joblib>=1.2
flask>=2.3.0
graphene>=3.3
graphql-core>=3.2.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0